serialisation with ox
we were serialising schedule object with YAML, but
ActiveSupport::TimeWithZone doesn’t play nice with it, since when
converting it to YAML it will be returned in UTC and lose time_zone
information. Tried to use Marshal to serialise it but couldn’t make it
to work, due to encoding problems when trying to do to_json on it.

So because of all this we will be using ox.